+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # AdManagerExtractor changelog
 
-## v0.0.2
+## v0.0.3
+
+- added support for `dimensionAttributes` and `adUnitView`
+- improved log output and error messages
+- default currency (CZK) is set only if some Ad Exchange metric is used
+
+## v0.0.2 (2020-10-22)
 
 - added config validation and better error handling